Crate flux_errors

Structs

Errors

Convenience struct implementing [ErrorEmitter] and [ErrorCollector]

FluxSession

ErrorGuaranteed

Useful type to use with Result<> indicate that an error has already been reported to the user, so no need to continue checking.

Constants

E0999

Functions

emitter 🔒