Crate flux_errors

Structs

ErrorGuaranteed

Useful type to use with Result<> indicate that an error has already been reported to the user, so no need to continue checking.

Errors

Convience struct implementing [ErrorEmitter] and [ErrorCollector]

FluxSession

Constants

E0999

Functions

emitter 🔒